Added deletion of bookmark list

+++ modules/nexus/app/services/NexusService.scala
@@ -348,7 +348,7 @@ class NexusService @Inject()(wSClient: WSClient, config:ConfigurationService)(im
     }
   }
 
-  def deprecateInstance(nexusEndpoint: String, nexusPath: NexusPath, id: String, revision: Int, token: String): Future[WSResponse] = {
+  def deprecateInstance(nexusEndpoint: String, nexusPath: NexusPath, id: String, revision: Long, token: String): Future[WSResponse] = {
     val instanceUrl = s"${nexusEndpoint}/v0/data/${nexusPath.org}/${nexusPath.domain}/${nexusPath.schema.toLowerCase}/${nexusPath.version}/${id}?rev=$revision"
     wSClient.url(instanceUrl).withHttpHeaders("Authorization" -> token).delete()
   }
